# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 趣聚 QuJu API
  version: 1.0.0
  description: >
    趣聚平台统一 API 契约（前后端唯一事实来源）。
    约定：所有业务接口返回统一信封 {code,message,data}，code=0 成功；列表用 page(从1)/size + PageResult；
    URL kebab-case 复数、JSON camelCase、前缀 /v1；JWT Bearer 鉴权；时间 ISO-8601(UTC)。
    枚举见 enums.md，错误码见 error-codes.md。改动须先写 ADR + 平台核心组评审。
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /teams:
    get:
      tags: [teams]
      summary: 发现/搜索小队（名称/标签）
      parameters: [{ name: keyword, in: query, schema: { type: string } }, { name: tag, in: query, schema: { type: string } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    post:
      tags: [teams]
      summary: 创建小队（创建者=队长）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TeamUpsertRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /admin/teams:
    get: { tags: [admin], summary: 查询小队, parameters: [{ name: keyword, in: query, schema: { type: string } }, { name: status, in: query, schema: { type: string } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /admin/teams/{id}/suspend:
    post: { tags: [admin], summary: 停用小队(必填原因), parameters: [{ $ref: '#/components/parameters/PathId' }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string } } } } } }, responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /admin/teams/{id}/restore:
    post: { tags: [admin], summary: 恢复小队, parameters: [{ $ref: '#/components/parameters/PathId' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /admin/reports:
    get: { tags: [admin], summary: 举报记录, parameters: [{ name: status, in: query, schema: { type: string } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /admin/reports/{id}/handle:
    post:
      tags: [admin]
      summary: 处理举报(驳回/标记已处理/处理并下架停用目标)
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [action], properties: { action: { type: string, enum: [DISMISS, RESOLVE, TAKEDOWN] }, reason: { type: string, description: 'action=TAKEDOWN 时必填' } } } } } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Size: { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    PathId: { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
  responses:
    Ok:
      description: 成功（data 按接口语义）
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { description: 业务数据 }
...
    TeamJoinType: { type: string, enum: [PUBLIC, APPROVAL] }
...
    TeamUpsertRequest:
      type: object
      required: [name, joinType]
      properties:
        name: { type: string }
        intro: { type: string }
        avatar: { type: string }
        tags: { type: array, items: { type: string } }
        joinType: { $ref: '#/components/schemas/TeamJoinType' }
        capacity: { type: integer }